requirements.txt
Django<2
python-decouple
mechanize
BeautifulSoup
cachetools
django-after-response
